+++ src/main.rs
@@ -8,17 +8,23 @@ mod parser;
 mod print;
 mod util;
 
-use crate::config::{Config, LanguageSettings, ParserSettings};
-use crate::document::Document;
-use crate::parser::code::{CodeParser, RevCodeBlock};
-use crate::util::{modify_path, Fallible, JoinExt};
-use clap::{crate_version, App, Arg, SubCommand};
-use std::collections::hash_map::Entry::{Occupied, Vacant};
-use std::collections::{HashMap, HashSet};
+use std::collections::{
+    hash_map::Entry::{Occupied, Vacant},
+    HashMap, HashSet,
+};
 use std::env::set_current_dir;
 use std::fs::{self, File};
 use std::io::Write;
 use std::path::{Path, PathBuf};
+
+use clap::{crate_version, App, Arg, SubCommand};
+
+use crate::{
+    config::Config,
+    document::Document,
+    parser::code::{CodeParser, RevCodeBlock},
+    util::{Fallible, JoinExt},
+};
 
 fn main() {
     std::process::exit(match run() {
@@ -249,7 +255,7 @@ fn process_inputs_reverse(
 
                 let parser = config.parser.default_language(code_type);
 
-                if let Err(error) = compile_all_reverse(
+                if let Err(error) = compile_reverse::compile_all(
                     &parser,
                     doc_dir,
                     code_dir,
@@ -471,7 +477,7 @@ fn copy_files(
                 }
             };
             if file.is_file() {
-                let out_path = path.map_or(file.clone(), |path| modify_path(&file, &path));
+                let out_path = path.map_or(file.clone(), |path| util::modify_path(&file, &path));
                 match track_copy_dest.entry(out_path.clone()) {
                     Occupied(entry) => {
                         return Err(format!(
@@ -505,153 +511,5 @@ fn copy_files(
             }
         }
     }
-    Ok(())
-}
-
-fn transclude_dry_run(
-    parser: &ParserSettings,
-    file_name: &Path,
-    code_dir: Option<&Path>,
-    entrypoint: Option<&str>,
-    language: Option<&str>,
-    documents: &mut HashMap<PathBuf, Document>,
-    track_code_files: &mut HashSet<PathBuf>,
-) -> Fallible<(Document, Vec<PathBuf>)> {
-    let source_main = util::read_file(&file_name)?;
-    let (document, mut links) = parse::parse(&source_main, &file_name, true, parser)?;
-
-    let transclusions = document.transclusions();
-
-    let mut trans_so_far = HashSet::new();
-    for trans in transclusions {
-        if !trans_so_far.contains(trans.file()) {
-            let (doc, sub_links) = transclude_dry_run(
-                parser,
-                trans.file(),
-                code_dir,
-                entrypoint,
-                language,
-                documents,
-                track_code_files,
-            )?;
-
-            compile_reverse(
-                parser,
-                &doc,
-                code_dir,
-                trans.file(),
-                entrypoint,
-                language,
-                track_code_files,
-            )?;
-
-            links.extend(sub_links.into_iter());
-            documents.insert(trans.file().clone(), doc);
-            trans_so_far.insert(trans.file().clone());
-        } else {
-            return Err(format!("Multiple transclusions of {}", trans.file().display()).into());
-        }
-    }
-
-    Ok((document, links))
-}
-
-#[allow(clippy::too_many_arguments)]
-fn compile_all_reverse(
-    parser: &ParserSettings,
-    doc_dir: Option<&Path>,
-    code_dir: Option<&Path>,
-    file_name: &Path,
-    entrypoint: Option<&str>,
-    language: Option<&str>,
-    settings: &HashMap<String, LanguageSettings>,
-    track_input_files: &mut HashSet<PathBuf>,
-    track_code_files: &mut HashSet<PathBuf>,
-    documents: &mut HashMap<PathBuf, Document>,
-) -> Fallible {
-    if !track_input_files.contains(file_name) {
-        let (mut document, links) = transclude_dry_run(
-            parser,
-            file_name,
-            code_dir,
-            entrypoint,
-            language,
-            documents,
-            track_code_files,
-        )?;
-
-        let file_str = file_name.to_str().unwrap();
-        document.set_source(file_str);
-
-        compile_reverse(
-            parser,
-            &document,
-            code_dir,
-            file_name,
-            entrypoint,
-            language,
-            track_code_files,
-        )?;
-
-        documents.insert(file_name.to_owned(), document);
-
-        track_input_files.insert(file_name.to_owned());
-
-        for file in links {
-            if !track_input_files.contains(&file) {
-                compile_all_reverse(
-                    parser,
-                    doc_dir,
-                    code_dir,
-                    &file,
-                    entrypoint,
-                    language,
-                    settings,
-                    track_input_files,
-                    track_code_files,
-                    documents,
-                )?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn compile_reverse(
-    parser: &ParserSettings,
-    document: &Document,
-    code_dir: Option<&Path>,
-    file_name: &Path,
-    entrypoint: Option<&str>,
-    language: Option<&str>,
-    track_code_files: &mut HashSet<PathBuf>,
-) -> Fallible {
-    eprintln!("Compiling file {}", file_name.display());
-
-    let mut entries = document.entry_points(parser, language);
-
-    let file_name_without_ext = file_name.with_extension("");
-    entries.insert(
-        entrypoint,
-        (&file_name_without_ext, Some(PathBuf::from(file_name))),
-    );
-
-    for (_entrypoint, (sub_file_name, _sub_source_file)) in entries {
-        match code_dir {
-            Some(code_dir) => {
-                let mut file_path = code_dir.to_owned();
-                file_path.push(sub_file_name);
-                if let Some(language) = language {
-                    file_path.set_extension(language);
-                }
-
-                track_code_files.insert(file_path);
-            }
-            None => eprintln!("WARNING: Missing output location for code, skipping code output."),
-        }
-    }
-
     Ok(())
 }
